requirements.txt
Django==1.6.5
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-guardian==1.2.0
django-toolbelt==0.0.1
django-userena==1.3.0
djrill==1.1.1
easy-thumbnails==2.0.1
gunicorn==19.0.0
html2text==2014.4.5
oauthlib==0.6.3
psycopg2==2.5.3
pystache==0.5.4
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.2
static==1.0.2
wsgiref==0.1.2
